third_party/blink/web_tests/external/wpt/editing/run/insertimage-expected.txt

This is a testharness.js-based test.
Found 30 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["insertimage","/img/lion.svg"]] "<span>foo[</span><span>]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span>foo<img src=\\"/img/lion.svg\\"></span><span>bar</span>" but got "<span>foo</span><img src=\\"/img/lion.svg\\"><span>bar</span>"
[FAIL] [["insertimage",""]] "foo[bar]baz": execCommand("insertimage", false, "") return value
  assert_equals: expected false but got true
[FAIL] [["insertimage",""]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<img>baz"
[FAIL] [["insertimage","/img/lion.svg"]] "foo<span style=color:#aBcDeF>[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"color:rgb(171, 205, 239)\\"><img src=\\"/img/lion.svg\\"></span>baz" but got "foo<img src=\\"/img/lion.svg\\">baz"
[FAIL] [["insertimage","/img/lion.svg"]] "foo<span style=color:#aBcDeF>{bar}</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"color:rgb(171, 205, 239)\\"><img src=\\"/img/lion.svg\\"></span>baz" but got "foo<img src=\\"/img/lion.svg\\">baz"
[FAIL] [["insertimage","/img/lion.svg"]] "foo{<span style=color:#aBcDeF>bar</span>}baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"color:rgb(171, 205, 239)\\"><img src=\\"/img/lion.svg\\"></span>baz" but got "foo<img src=\\"/img/lion.svg\\">baz"
[FAIL] [["stylewithcss","true"],["insertimage","/img/lion.svg"]] "[foo<span style=color:#aBcDeF>bar]</span>baz" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["insertimage","/img/lion.svg"]] "foo<span style=color:#aBcDeF>[bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"color:rgb(171, 205, 239)\\"><img src=\\"/img/lion.svg\\"></span>" but got "foo<img src=\\"/img/lion.svg\\">"
[FAIL] [["insertimage","/img/lion.svg"]] "foo<span style=color:#aBcDeF>{bar</span>baz}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"color:rgb(171, 205, 239)\\"><img src=\\"/img/lion.svg\\"></span>" but got "foo<img src=\\"/img/lion.svg\\">"
[FAIL] [["stylewithcss","true"],["insertimage","/img/lion.svg"]] "foo<span style=color:#aBcDeF>[bar</span><span style=color:#fEdCbA>baz]</span>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"color:rgb(171, 205, 239)\\"><img src=\\"/img/lion.svg\\"></span>quz" but got "foo<img src=\\"/img/lion.svg\\">quz"
[FAIL] [["stylewithcss","false"],["insertimage","/img/lion.svg"]] "foo<span style=color:#aBcDeF>[bar</span><span style=color:#fEdCbA>baz]</span>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"color:rgb(171, 205, 239)\\"><img src=\\"/img/lion.svg\\"></span>quz" but got "foo<img src=\\"/img/lion.svg\\">quz"
[FAIL] [["insertimage","/img/lion.svg"]] "foo<b>[bar]</b>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<b><img src=\\"/img/lion.svg\\"></b>baz" but got "foo<img src=\\"/img/lion.svg\\">baz"
[FAIL] [["insertimage","/img/lion.svg"]] "foo<b>{bar}</b>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<b><img src=\\"/img/lion.svg\\"></b>baz" but got "foo<img src=\\"/img/lion.svg\\">baz"
[FAIL] [["insertimage","/img/lion.svg"]] "foo{<b>bar</b>}baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<b><img src=\\"/img/lion.svg\\"></b>baz" but got "foo<img src=\\"/img/lion.svg\\">baz"
[FAIL] [["insertimage","/img/lion.svg"]] "foo<span>[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span><img src=\\"/img/lion.svg\\"></span>baz" but got "foo<img src=\\"/img/lion.svg\\">baz"
[FAIL] [["insertimage","/img/lion.svg"]] "foo<span>{bar}</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span><img src=\\"/img/lion.svg\\"></span>baz" but got "foo<img src=\\"/img/lion.svg\\">baz"
[FAIL] [["insertimage","/img/lion.svg"]] "foo{<span>bar</span>}baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span><img src=\\"/img/lion.svg\\"></span>baz" but got "foo<img src=\\"/img/lion.svg\\">baz"
[FAIL] [["insertimage","/img/lion.svg"]] "<b>foo[bar</b><i>baz]quz</i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foo<img src=\\"/img/lion.svg\\"></b><i>quz</i>" but got "<b>foo</b><img src=\\"/img/lion.svg\\"><i>quz</i>"
[FAIL] [["defaultparagraphseparator","div"],["insertimage","/img/lion.svg"]] "<p>foo</p>{<p>bar</p>}<p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><img src=\\"/img/lion.svg\\"><p>baz</p>" but got "<p>foo</p><p><img src=\\"/img/lion.svg\\">baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertimage","/img/lion.svg"]] "<p>foo</p>{<p>bar</p>}<p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><img src=\\"/img/lion.svg\\"><p>baz</p>" but got "<p>foo</p><p><img src=\\"/img/lion.svg\\">baz</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertimage","/img/lion.svg"]] "<p>foo[bar<p style=color:blue>baz]quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\"><font color=\\"#0000ff\\">quz</font></p>" but got "<p>foo<img src=\\"/img/lion.svg\\"><span style=\\"color:rgb(0, 0, 255)\\">quz</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertimage","/img/lion.svg"]] "<p>foo[bar<p style=color:blue>baz]quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\"><font color=\\"#0000ff\\">quz</font></p>" but got "<p>foo<img src=\\"/img/lion.svg\\"><span style=\\"color:rgb(0, 0, 255)\\">quz</span></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertimage","/img/lion.svg"]] "<div><p>foo<p>[bar<p>baz]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p>foo</p><p><img src=\\"/img/lion.svg\\"></p></div>" but got "<p>foo</p><p><img src=\\"/img/lion.svg\\"></p>"
[FAIL] [["defaultparagraphseparator","p"],["insertimage","/img/lion.svg"]] "<div><p>foo<p>[bar<p>baz]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p>foo</p><p><img src=\\"/img/lion.svg\\"></p></div>" but got "<p>foo</p><p><img src=\\"/img/lion.svg\\"></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertimage","/img/lion.svg"]] "<p>foo[</p><p>]bar<br>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\">bar<br>baz</p>" but got "<p>foo<img src=\\"/img/lion.svg\\">bar</p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertimage","/img/lion.svg"]] "<p>foo[</p><p>]bar<br>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\">bar<br>baz</p>" but got "<p>foo<img src=\\"/img/lion.svg\\">bar</p><p>baz</p>"
[FAIL] [["insertimage","/img/lion.svg"]] "<p>foo[</p>]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\">bar</p>" but got "<p>foo<img src=\\"/img/lion.svg\\"></p>bar"
[FAIL] [["insertimage","/img/lion.svg"]] "<p>foo[</p>]bar<br>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\">bar</p>baz" but got "<p>foo<img src=\\"/img/lion.svg\\"></p>bar<br>baz"
[FAIL] [["insertimage","/img/lion.svg"]] "<p>foo[</p>]bar<p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\">bar</p><p>baz</p>" but got "<p>foo<img src=\\"/img/lion.svg\\"></p>bar<p>baz</p>"
[FAIL] [["insertimage","/img/lion.svg"]] "<div><p>foo</p>bar[</div>]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p>foo</p>bar<img src=\\"/img/lion.svg\\">baz</div>" but got "<div><p>foo</p>bar<img src=\\"/img/lion.svg\\"></div>baz"
Harness: the test ran to completion.
